+++ internal/dspscraper/types.go
@@ -2,5 +2,5 @@ package dspscraper
 
 type ComputedRecipeRequest struct {
 	Name string  `json:"name" validate:"required"`
-	Rate float32 `json:"rate" validate:"required"`
+	Rate float64 `json:"rate" validate:"required"`
 }
